pyproject.toml
[project]
name = "django-rest-framework"
version = "0.1.0"
description = "Django REST Framework Tutorial"
authors = [
    {name = "iBrokeTheCode",email = "alexrtc.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-silk = "^5.3.2"

